- Refactor Command#introspection
