+++ lib/bashly/script/command.rb
@@ -2,6 +2,12 @@ module Bashly
   module Script
     class Command < Base
       include Completions::Command
+      include Introspection::Arguments
+      include Introspection::Commands
+      include Introspection::Dependencies
+      include Introspection::EnvironmentVariables
+      include Introspection::Examples
+      include Introspection::Flags
 
       class << self
         def option_keys
@@ -9,7 +15,7 @@ module Bashly
             alias args catch_all commands completions
             default dependencies environment_variables examples
             extensible expose filename filters flags
-            footer function group help name
+            footer function group help name 
             private version
           ]
         end
@@ -39,15 +45,6 @@ module Bashly
         options['alias'].is_a?(String) ? [options['alias']] : options['alias']
       end
 
-      # Returns an array of Arguments
-      def args
-        return [] unless options['args']
-
-        options['args'].map do |options|
-          Argument.new options
-        end
-      end
-
       # Returns a string suitable to be a headline
       def caption_string
         help.empty? ? full_name : "#{full_name} - #{summary}"
@@ -58,121 +55,10 @@ module Bashly
         @catch_all ||= CatchAll.from_config options['catch_all']
       end
 
-      # Returns a full list of the Command names and aliases combined
-      def command_aliases
-        commands.map(&:aliases).flatten
-      end
-
-      # Returns a data structure for displaying subcommands help
-      def command_help_data
-        result = {}
-
-        public_commands.each do |command|
-          result[command.group_string] ||= {}
-          result[command.group_string][command.name] = { summary: command.summary_string }
-          next unless command.expose
-
-          command.public_commands.each do |subcommand|
-            result[command.group_string]["#{command.name} #{subcommand.name}"] = {
-              summary:   subcommand.summary_string,
-              help_only: command.expose != 'always',
-            }
-          end
-        end
-
-        result
-      end
-
-      # Returns only the names of the Commands
-      def command_names
-        commands.map(&:name)
-      end
-
-      # Returns an array of the Commands
-      def commands
-        return [] unless options['commands']
-
-        options['commands'].map do |options|
-          result = Command.new options
-          result.parents = parents + [name]
-          result.parent_command = self
-          result
-        end
-      end
-
-      # Returns a flat array containing all the commands in this tree.
-      # This includes children + grandchildren (recursive), and may include self
-      def deep_commands(include_self: false)
-        result = []
-        result << self if include_self
-        commands.each do |command|
-          result << command
-          if command.commands.any?
-            result += command.deep_commands
-          end
-        end
-        result
-      end
-
-      # If any of this command's subcommands has the default option set to
-      # true, this default command will be returned, nil otherwise.
-      def default_command
-        commands.find(&:default)
-      end
-
-      # Returns an array of all the default Args
-      def default_args
-        args.select(&:default)
-      end
-
-      # Returns an array of all the default Environment Variables
-      def default_environment_variables
-        environment_variables.select(&:default)
-      end
-
-      # Returns an array of all the default Flags
-      def default_flags
-        flags.select(&:default)
-      end
-
-      # Returns an array of Dependency objects
-      def dependencies
-        return [] unless options['dependencies']
-
-        @dependencies ||= options['dependencies'].map do |key, value|
-          Dependency.from_config key, value
-        end
-      end
-
-      # Returns an array of EnvironmentVariable objects
-      def environment_variables
-        return [] unless options['environment_variables']
-
-        options['environment_variables'].map do |options|
-          EnvironmentVariable.new options
-        end
-      end
-
-      # Returns an array of examples
-      def examples
-        return nil unless options['examples']
-
-        options['examples'].is_a?(Array) ? options['examples'] : [options['examples']]
-      end
-
       # Returns the filename that is expected to hold the user code for this
       # command
       def filename
         options['filename'] || implicit_filename
-      end
-
-      # Returns an array of Flags
-      def flags
-        return [] unless options['flags']
-
-        options['flags'].map do |options|
-          Flag.new options
-        end
       end
 
       # Returns a unique name, suitable to be used in a bash function
@@ -186,12 +72,6 @@ module Bashly
         parents.any? ? (parents + [name]).join(' ') : name
       end
 
-      # Returns true if this command's flags should be considered as gloal
-      # flags, and cascade to subcommands
-      def global_flags?
-        flags.any? and commands.any?
-      end
-
       # Returns the string for the group caption
       def group_string
         if group
@@ -199,23 +79,6 @@ module Bashly
         else
           strings[:commands]
         end
-      end
-
-      # Returns subcommands by group
-      def grouped_commands
-        result = {}
-
-        public_commands.each do |command|
-          result[command.group_string] ||= []
-          result[command.group_string] << command
-          next unless command.expose
-
-          command.public_commands.each do |subcommand|
-            result[command.group_string] << subcommand
-          end
-        end
-
-        result
       end
 
       # Returns true if this command, or any subcommand (deep) as any arg or
@@ -246,58 +109,9 @@ module Bashly
         @parents ||= []
       end
 
-      # Returns only commands that are not private
-      def public_commands
-        commands.reject(&:private)
-      end
-
-      # Returns a full list of the public Command names and aliases combined
-      def public_command_aliases
-        public_commands.map(&:aliases).flatten
-      end
-
-      # Returns only environment variables that are not private
-      def public_environment_variables
-        environment_variables.reject(&:private)
-      end
-
-      # Returns only flags that are not private
-      def public_flags
-        flags.reject(&:private)
-      end
-
-      # Returns true if one of the args is repeatable
-      def repeatable_arg_exist?
-        args.any?(&:repeatable)
-      end
-
-      # Returns an array of all the required Arguments
-      def required_args
-        args.select(&:required)
-      end
-
-      # Returns an array of all the required EnvironmentVariables
-      def required_environment_variables
-        environment_variables.select(&:required)
-      end
-
-      # Returns an array of all the required Flags
-      def required_flags
-        flags.select(&:required)
-      end
-
-      def needy_flags
-        flags.select(&:needs)
-      end
-
       # Returns true if this is the root command (no parents)
       def root_command?
         parents.empty?
-      end
-
-      # Returns true if one of the flags matches the provided short code
-      def short_flag_exist?(flag)
-        flags.any? { |f| f.short == flag }
       end
 
       # Returns the summary string
@@ -326,11 +140,8 @@ module Bashly
         result.compact.join ' '
       end
 
-      # Returns an array of args usage_string for the command's usage_string
-      def usage_string_args
-        args.map(&:usage_string)
-      end
-
+      # Returns a base usage string that considers whether this command is the
+      # default, and if it has any parents. Used internally by `usage_string`.
       def base_usage_pattern
         usage_pattern = default ? "[#{name}]" : name
         parents.any? ? (parents + [usage_pattern]).join(' ') : usage_pattern
@@ -346,21 +157,6 @@ module Bashly
       # Returns a mixed array of Argument and Flag objects that have validations
       def validatables
         @validatables ||= args.select(&:validate) + flags.select(&:validate)
-      end
-
-      # Returns an array of all the args with a whitelist
-      def whitelisted_args
-        args.select(&:allowed)
-      end
-
-      # Returns an array of all the environment_variables with a whitelist arg
-      def whitelisted_environment_variables
-        environment_variables.select(&:allowed)
-      end
-
-      # Returns an array of all the flags with a whitelist arg
-      def whitelisted_flags
-        flags.select(&:allowed)
       end
 
     private
