+++ spec/approvals/examples/needs
@@ -32,9 +32,9 @@ Options:
     Show version number
 
 + ./cli --add deploy
---add requires --command
+--add needs --command
 + ./cli --add deploy --command 'git push'
---add requires --target
+--add needs --target
 + ./cli --add deploy --command 'git push' --target local
 # this file is located in 'src/root_command.sh'
 # you can edit it freely and regenerate (it will not be overwritten)
