+++ spec/bashly/integration/bash_spec.rb
@@ -1,5 +1,5 @@
 describe 'bash' do
-  context 'when bash version is < 4' do
+  context 'when bash version is < 4', :slow do
     before { system 'docker pull bash:3 >/dev/null' }
 
     it 'errors gracefully' do
